fix: SubscriptionStart & Benefits thankyou components moved to their own file
